Lógica de controller

+++ services/product-service/Controllers/ProductController.cs
@@ -61,36 +61,38 @@ namespace product_service.Controllers
             return _response;
         }
 
-        /* [HttpPost]
+        [HttpPost]
         [Authorize(Roles = "ADMIN")]
-        public ResponseDto Post(ProductDto ProductDto)
+        public ResponseDto Post(ProductDto productDto)
         {
             try
             {
-                Product product = _mapper.Map<Product>(ProductDto);
+                Product product = _mapper.Map<Product>(productDto);
 
-                _db.Products.Add(product);
-                _db.SaveChanges();
+                _context.Products.Add(product);
+                _context.SaveChanges();
 
-                if (ProductDto.Image != null)
+                if (productDto.Image != null)
                 {
 
-                    string fileName = product.ProductId + Path.GetExtension(ProductDto.Image.FileName);
+                    string fileName = product.ProductId + Path.GetExtension(productDto.Image.FileName);
                     string filePath = @"wwwroot\ProductImages\" + fileName;
 
                     var directoryLocation = Path.Combine(Directory.GetCurrentDirectory(), filePath);
-                    FileInfo file = new FileInfo(directoryLocation);
-                    if (file.Exists)
-                    {
-                        file.Delete();
-                    }
+
+                    FileInfo file = new(directoryLocation);
+
+                    if (file.Exists) file.Delete();
 
                     var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
+
                     using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
                     {
-                        ProductDto.Image.CopyTo(fileStream);
+                        productDto.Image.CopyTo(fileStream);
                     }
+
                     var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
+
                     product.ImageUrl = baseUrl + "/ProductImages/" + fileName;
                     product.ImageLocalPath = filePath;
                 }
@@ -98,13 +100,15 @@ namespace product_service.Controllers
                 {
                     product.ImageUrl = "https://placehold.co/600x400";
                 }
-                _db.Products.Update(product);
-                _db.SaveChanges();
+
+                _context.Products.Update(product);
+                _context.SaveChanges();
+
                 _response.Result = _mapper.Map<ProductDto>(product);
             }
             catch (Exception ex)
             {
-                _response.IsSuccess = false;
+                _response.Success = false;
                 _response.Message = ex.Message;
             }
 
@@ -114,47 +118,51 @@ namespace product_service.Controllers
 
         [HttpPut]
         [Authorize(Roles = "ADMIN")]
-        public ResponseDto Put(ProductDto ProductDto)
+        public ResponseDto Put(ProductDto productDto)
         {
             try
             {
-                Product product = _mapper.Map<Product>(ProductDto);
+                Product product = _mapper.Map<Product>(productDto);
 
-                if (ProductDto.Image != null)
+                if (productDto.Image != null)
                 {
                     if (!string.IsNullOrEmpty(product.ImageLocalPath))
                     {
                         var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), product.ImageLocalPath);
+
                         FileInfo file = new FileInfo(oldFilePathDirectory);
-                        if (file.Exists)
-                        {
-                            file.Delete();
-                        }
+
+                        if (file.Exists) file.Delete();
                     }
 
-                    string fileName = product.ProductId + Path.GetExtension(ProductDto.Image.FileName);
+                    string fileName = product.ProductId + Path.GetExtension(productDto.Image.FileName);
                     string filePath = @"wwwroot\ProductImages\" + fileName;
+
                     var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
+
                     using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
                     {
-                        ProductDto.Image.CopyTo(fileStream);
+                        productDto.Image.CopyTo(fileStream);
                     }
+
                     var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
+
                     product.ImageUrl = baseUrl + "/ProductImages/" + fileName;
                     product.ImageLocalPath = filePath;
                 }
 
 
-                _db.Products.Update(product);
-                _db.SaveChanges();
+                _context.Products.Update(product);
+                _context.SaveChanges();
 
                 _response.Result = _mapper.Map<ProductDto>(product);
             }
             catch (Exception ex)
             {
-                _response.IsSuccess = false;
+                _response.Success = false;
                 _response.Message = ex.Message;
             }
+
             return _response;
         }
 
@@ -165,25 +173,27 @@ namespace product_service.Controllers
         {
             try
             {
-                Product obj = _db.Products.First(u => u.ProductId == id);
+                Product obj = _context.Products.First(u => u.ProductId == id);
+
                 if (!string.IsNullOrEmpty(obj.ImageLocalPath))
                 {
                     var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), obj.ImageLocalPath);
-                    FileInfo file = new FileInfo(oldFilePathDirectory);
-                    if (file.Exists)
-                    {
-                        file.Delete();
-                    }
+
+                    FileInfo file = new(oldFilePathDirectory);
+
+                    if (file.Exists) file.Delete();
                 }
-                _db.Products.Remove(obj);
-                _db.SaveChanges();
+
+                _context.Products.Remove(obj);
+                _context.SaveChanges();
             }
             catch (Exception ex)
             {
-                _response.IsSuccess = false;
+                _response.Success = false;
                 _response.Message = ex.Message;
             }
+
             return _response;
-        } */
+        }
     }
 }
